Could you please add my school to the list?
Could you kindly add our school "Bertha Parvis Institute of Technology" to the JetBrains educational programme list so we can make use of the free licences? We promise that all licences will be used exclusively for legitimate educational activities with our students and staff. We’d be incredibly grateful for your approval.Many thanks in advance!